rename crypto.js
