+++ packages/message-encryption/src/crypto/index.ts
@@ -1,3 +1,3 @@
-export * from "./crypto.js";
+export * from "./utils.js";
 export * as ecies from "./ecies.js";
 export * as symmetric from "./symmetric.js";
